+++ astropy/time/tests/test_basic.py
@@ -5,17 +5,22 @@
 import copy
 import functools
 import sys
-
-from datetime import datetime, tzinfo, timedelta
+import datetime
 
 import numpy as np
 
 from ...tests.helper import pytest, catch_warnings
 from ...extern import six
 from ...utils import isiterable
-from .. import Time, ScaleValueError, erfa_time, TIME_SCALES, TimeString
+from .. import (Time, ScaleValueError, erfa_time, TIME_SCALES, TimeString,
+                TimezoneInfo)
 from ...coordinates import EarthLocation
-
+from ... import units as u
+try:
+    import pytz
+    HAS_PYTZ = True
+except ImportError:
+    HAS_PYTZ = False
 
 allclose_jd = functools.partial(np.allclose, rtol=2. ** -52, atol=0)
 allclose_jd2 = functools.partial(np.allclose, rtol=2. ** -52,
@@ -24,6 +29,8 @@ allclose_sec = functools.partial(np.allclose, rtol=2. ** -52,
                                  atol=2. ** -52 * 24 * 3600)  # 20 ps atol
 allclose_year = functools.partial(np.allclose, rtol=2. ** -52,
                                   atol=0.)  # 14 microsec at current epoch
+
+
 
 
 class TestBasic():
@@ -199,7 +206,7 @@ class TestBasic():
         assert allclose_sec(t.unix, 1262304000.0)
         assert allclose_sec(t.cxcsec, 378691266.184)
         assert allclose_sec(t.gps, 946339215.0)
-        assert t.datetime == datetime(2010, 1, 1)
+        assert t.datetime == datetime.datetime(2010, 1, 1)
 
     def test_precision(self):
         """Set the output precision which is used for some formats.  This is
@@ -342,7 +349,7 @@ class TestBasic():
         Time(0.0, 51544.0333981, format='mjd', scale='tai')
         Time('2000:001:12:23:34.0', format='yday', scale='tai')
         Time('2000:001:12:23:34.0Z', format='yday', scale='utc')
-        dt = datetime(2000, 1, 2, 3, 4, 5, 123456)
+        dt = datetime.datetime(2000, 1, 2, 3, 4, 5, 123456)
         Time(dt, format='datetime', scale='tai')
         Time([dt, dt], format='datetime', scale='tai')
 
@@ -351,8 +358,8 @@ class TestBasic():
         Test datetime format, including guessing the format from the input type
         by not providing the format keyword to Time.
         """
-        dt = datetime(2000, 1, 2, 3, 4, 5, 123456)
-        dt2 = datetime(2001, 1, 1)
+        dt = datetime.datetime(2000, 1, 2, 3, 4, 5, 123456)
+        dt2 = datetime.datetime(2001, 1, 1)
         t = Time(dt, scale='utc', precision=9)
         assert t.iso == '2000-01-02 03:04:05.123456000'
         assert t.datetime == dt
@@ -364,7 +371,7 @@ class TestBasic():
         assert np.all(t.value == [dt, dt2])
 
         t = Time('2000-01-01 01:01:01.123456789', scale='tai')
-        assert t.datetime == datetime(2000, 1, 1, 1, 1, 1, 123457)
+        assert t.datetime == datetime.datetime(2000, 1, 1, 1, 1, 1, 123457)
 
         # broadcasting
         dt3 = (dt + (dt2-dt)*np.arange(12)).reshape(4, 3)
@@ -780,7 +787,7 @@ def test_now():
     Tests creating a Time object with the `now` class method.
     """
 
-    now = datetime.utcnow()
+    now = datetime.datetime.utcnow()
     t = Time.now()
 
     assert t.format == 'datetime'
@@ -896,13 +903,13 @@ def test_datetime_tzinfo():
     """
     Test #3160 that time zone info in datetime objects is respected.
     """
-    class TZm6(tzinfo):
+    class TZm6(datetime.tzinfo):
         def utcoffset(self, dt):
-            return timedelta(hours=-6)
+            return datetime.timedelta(hours=-6)
 
-    d = datetime(2002, 1, 2, 10, 3, 4, tzinfo=TZm6())
+    d = datetime.datetime(2002, 1, 2, 10, 3, 4, tzinfo=TZm6())
     t = Time(d)
-    assert t.value == datetime(2002, 1, 2, 16, 3, 4)
+    assert t.value == datetime.datetime(2002, 1, 2, 16, 3, 4)
 
 def test_subfmts_regex():
     """
@@ -924,7 +931,7 @@ def test_set_format_basic():
     for format, value in (('jd', 2451577.5),
                           ('mjd', 51577.0),
                           ('cxcsec', 65923200.0),
-                          ('datetime', datetime(2000, 2, 3, 0, 0)),
+                          ('datetime', datetime.datetime(2000, 2, 3, 0, 0)),
                           ('iso', '2000-02-03 00:00:00.000')):
         t = Time('+02000-02-03', format='fits')
         t0 = t.replicate()
@@ -999,3 +1006,43 @@ def test_isiterable():
     t2 = Time(['1999-01-01 00:00:00.123456789', '2010-01-01 00:00:00'],
               format='iso', scale='utc')
     assert isiterable(t2)
+
+def test_to_datetime():
+    tz = TimezoneInfo(utc_offset=-10*u.hour, tzname='US/Hawaii')
+    # The above lines produces a `datetime.tzinfo` object similar to:
+    #     tzinfo = pytz.timezone('US/Hawaii')
+    time = Time('2010-09-03 00:00:00')
+    tz_aware_datetime = time.to_datetime(tz)
+    assert tz_aware_datetime.time() == datetime.time(14, 0)
+    forced_to_astropy_time = Time(tz_aware_datetime)
+    assert tz.tzname(time.datetime) == tz_aware_datetime.tzname()
+    assert time == forced_to_astropy_time
+
+    # Test non-scalar time inputs:
+    time = Time(['2010-09-03 00:00:00', '2005-09-03 06:00:00',
+                 '1990-09-03 06:00:00'])
+    tz_aware_datetime = time.to_datetime(tz)
+    forced_to_astropy_time = Time(tz_aware_datetime)
+    for dt, tz_dt in zip(time.datetime, tz_aware_datetime):
+        assert tz.tzname(dt) == tz_dt.tzname()
+    assert np.all(time == forced_to_astropy_time)
+
+@pytest.mark.skipif('not HAS_PYTZ')
+def test_to_datetime_pytz():
+
+    tz = pytz.timezone('US/Hawaii')
+    time = Time('2010-09-03 00:00:00')
+    tz_aware_datetime = time.to_datetime(tz)
+    forced_to_astropy_time = Time(tz_aware_datetime)
+    assert tz_aware_datetime.time() == datetime.time(14, 0)
+    assert tz.tzname(time.datetime) == tz_aware_datetime.tzname()
+    assert time == forced_to_astropy_time
+
+    # Test non-scalar time inputs:
+    time = Time(['2010-09-03 00:00:00', '2005-09-03 06:00:00',
+                 '1990-09-03 06:00:00'])
+    tz_aware_datetime = time.to_datetime(tz)
+    forced_to_astropy_time = Time(tz_aware_datetime)
+    for dt, tz_dt in zip(time.datetime, tz_aware_datetime):
+        assert tz.tzname(dt) == tz_dt.tzname()
+    assert np.all(time == forced_to_astropy_time)